PhoneValidation/Backend/app_b86ed9a3-cfa5-4d79-a572-890b5a4e545f/requirements.txt
django
djangorestframework
django-cors-middleware
smsapi-client==2.4.2
